Hybrid Car Evaluation and Payment Calculator
An Application that allows the user to decide if buying a hybrid car is
worth the extra cost, based on the total cost of ownership over a 10
year period using .Net GUI Components, Object Oriented features, file
I/O and Exception Handling using C# .Net. Allows the user to choose one
car, enter a down payment amount, an annual interest rate and a loan
period and produce an amortization report for a car loan.